Move deploy helpers to submodules and simplify flow
