+++ crates/cloud/src/cloud_client_extensions.rs
@@ -6,13 +6,28 @@ use crate::CloudClientInterface;
 
 #[async_trait]
 pub trait CloudClientExt {
+    async fn add_revision_ref(&self, storage_id: &str, version: &str) -> anyhow::Result<()>;
+
     async fn get_app_id(&self, app_name: &str) -> Result<Option<Uuid>>;
     async fn get_revision_id(&self, app_id: Uuid, version: &str) -> Result<Uuid>;
     async fn get_channel_id(&self, app_id: Uuid, channel_name: &str) -> Result<Uuid>;
+
+    async fn set_key_values(
+        &self,
+        app_id: Uuid,
+        store_label: &str,
+        key_values: &[(String, String)],
+    ) -> Result<()>;
+    async fn set_variables(&self, app_id: Uuid, variables: &[(String, String)]) -> Result<()>;
 }
 
 #[async_trait]
 impl<T: CloudClientInterface> CloudClientExt for T {
+    async fn add_revision_ref(&self, storage_id: &str, version: &str) -> anyhow::Result<()> {
+        self.add_revision(storage_id.to_owned(), version.to_owned())
+            .await
+    }
+
     async fn get_app_id(&self, app_name: &str) -> Result<Option<Uuid>> {
         let apps_vm = self
             .list_apps(crate::DEFAULT_APPLIST_PAGE_SIZE, None)
@@ -72,5 +87,33 @@ impl<T: CloudClientInterface> CloudClientExt for T {
             app_id,
             channel_name,
         ))
+    }
+
+    async fn set_key_values(
+        &self,
+        app_id: Uuid,
+        store_label: &str,
+        key_values: &[(String, String)],
+    ) -> Result<()> {
+        for (key, value) in key_values {
+            self.add_key_value_pair(
+                app_id,
+                store_label.to_owned(),
+                key.to_owned(),
+                value.to_owned(),
+            )
+            .await
+            .with_context(|| format!("Problem creating key/value {}", key))?;
+        }
+        Ok(())
+    }
+
+    async fn set_variables(&self, app_id: Uuid, variables: &[(String, String)]) -> Result<()> {
+        for (name, value) in variables {
+            self.add_variable_pair(app_id, name.to_owned(), value.to_owned())
+                .await
+                .with_context(|| format!("Problem creating variable {}", name))?;
+        }
+        Ok(())
     }
 }
