+++ src/commands/deploy/cancellable.rs
@@ -1,0 +1,4 @@
+pub enum Cancellable<T> {
+    Cancelled,
+    Accepted(T),
+}
