+++ src/commands/deploy/login.rs
@@ -1,0 +1,171 @@
+use std::{
+    io::{self},
+    path::PathBuf,
+};
+
+use anyhow::{anyhow, bail, Context, Result};
+use chrono::{DateTime, Utc};
+use clap::Parser;
+use cloud::client::{Client as CloudClient, ConnectionConfig};
+use cloud::CloudClientInterface;
+use spin_common::sloth;
+use tokio::fs;
+use url::Url;
+
+use crate::commands::login::{LoginCommand, LoginConnection};
+
+pub async fn login_connection(deployment_env_id: Option<&str>) -> Result<LoginConnection> {
+    let path = config_file_path(deployment_env_id)?;
+
+    // log in if config.json does not exist or cannot be read
+    let data = match fs::read_to_string(path.clone()).await {
+        Ok(d) => d,
+        Err(e) if e.kind() == io::ErrorKind::NotFound => {
+            match deployment_env_id {
+                Some(name) => {
+                    // TODO: allow auto redirect to login preserving the name
+                    eprintln!("You have no instance saved as '{}'", name);
+                    eprintln!("Run `spin login --environment-name {}` to log in", name);
+                    std::process::exit(1);
+                }
+                None => {
+                    // log in, then read config
+                    // TODO: propagate deployment id (or bail if nondefault?)
+                    LoginCommand::parse_from(vec!["login"]).run().await?;
+                    fs::read_to_string(path.clone()).await?
+                }
+            }
+        }
+        Err(e) => {
+            bail!("Could not log in: {}", e);
+        }
+    };
+
+    let mut login_connection: LoginConnection = serde_json::from_str(&data)?;
+    let expired = match has_expired(&login_connection) {
+        Ok(val) => val,
+        Err(err) => {
+            eprintln!("{}\n", err);
+            eprintln!("Run `spin login` to log in again");
+            std::process::exit(1);
+        }
+    };
+
+    if expired {
+        // if we have a refresh token available, let's try to refresh the token
+        match login_connection.refresh_token {
+            Some(refresh_token) => {
+                // Only Cloud has support for refresh tokens
+                let connection_config = ConnectionConfig {
+                    url: login_connection.url.to_string(),
+                    insecure: login_connection.danger_accept_invalid_certs,
+                    token: login_connection.token.clone(),
+                };
+                let client = CloudClient::new(connection_config.clone());
+
+                match client
+                    .refresh_token(login_connection.token, refresh_token)
+                    .await
+                {
+                    Ok(token_info) => {
+                        login_connection.token = token_info.token;
+                        login_connection.refresh_token = Some(token_info.refresh_token);
+                        login_connection.expiration = Some(token_info.expiration);
+                        // save new token info
+                        let path = config_file_path(deployment_env_id)?;
+                        std::fs::write(path, serde_json::to_string_pretty(&login_connection)?)?;
+                    }
+                    Err(e) => {
+                        eprintln!("Failed to refresh token: {}", e);
+                        match deployment_env_id {
+                            Some(name) => {
+                                eprintln!(
+                                    "Run `spin login --environment-name {}` to log in again",
+                                    name
+                                );
+                            }
+                            None => {
+                                eprintln!("Run `spin login` to log in again");
+                            }
+                        }
+                        std::process::exit(1);
+                    }
+                }
+            }
+            None => {
+                // session has expired and we have no way to refresh the token - log back in
+                match deployment_env_id {
+                    Some(name) => {
+                        // TODO: allow auto redirect to login preserving the name
+                        eprintln!("Your login to this environment has expired");
+                        eprintln!(
+                            "Run `spin login --environment-name {}` to log in again",
+                            name
+                        );
+                        std::process::exit(1);
+                    }
+                    None => {
+                        LoginCommand::parse_from(vec!["login"]).run().await?;
+                        let new_data = fs::read_to_string(path.clone()).await.context(format!(
+                            "Cannot find spin config at {}",
+                            path.to_string_lossy()
+                        ))?;
+                        login_connection = serde_json::from_str(&new_data)?;
+                    }
+                }
+            }
+        }
+    }
+
+    let sloth_guard = sloth::warn_if_slothful(
+        2500,
+        format!("Checking status ({})\n", login_connection.url),
+    );
+    check_healthz(&login_connection.url).await?;
+    // Server has responded - we don't want to keep the sloth timer running.
+    drop(sloth_guard);
+
+    Ok(login_connection)
+}
+
+// TODO: unify with login
+fn config_file_path(deployment_env_id: Option<&str>) -> Result<PathBuf> {
+    let root = dirs::config_dir()
+        .context("Cannot find configuration directory")?
+        .join("fermyon");
+
+    let file_stem = match deployment_env_id {
+        None => "config",
+        Some(id) => id,
+    };
+    let file = format!("{}.json", file_stem);
+
+    let path = root.join(file);
+
+    Ok(path)
+}
+
+// Check if the token has expired.
+// If the expiration is None, assume the token has not expired
+fn has_expired(login_connection: &LoginConnection) -> Result<bool> {
+    match &login_connection.expiration {
+        Some(expiration) => match DateTime::parse_from_rfc3339(expiration) {
+            Ok(time) => Ok(Utc::now() > time),
+            Err(err) => Err(anyhow!(
+                "Failed to parse token expiration time '{}'. Error: {}",
+                expiration,
+                err
+            )),
+        },
+        None => Ok(false),
+    }
+}
+
+async fn check_healthz(base_url: &Url) -> Result<()> {
+    let healthz_url = base_url.join("healthz")?;
+    reqwest::get(healthz_url)
+        .await?
+        .error_for_status()
+        .with_context(|| format!("Server {} is unhealthy", base_url))?;
+    Ok(())
+}
